Move getTypeVisualizer to BaseContikiMoteType
ContikiMoteType displays a bit more information than
the native nodes, so add a target-hook to fill in those
parts.
